+++ .gitignore
@@ -1,2 +1,3 @@
 states
 states-test
+.idea
